flx::geolookup::models

Struct GeoData

pub struct GeoData {

    pub iso_code: Option<Box<str>>,
    pub name: Option<Box<str>>,
    pub region_iso_code: Option<Box<str>>,
    pub region_name: Option<Box<str>>,
    pub city_name: Option<Box<str>>,
    pub asn: Option<u32>,
    pub aso: Option<Box<str>>,
    pub ip_type: IpType,
    pub continent_code: Option<Box<str>>,
    pub continent_name: Option<Box<str>>,
    pub latitude: Option<f64>,
    pub longitude: Option<f64>,
    pub timezone: Option<Box<str>>,
    pub zip: Option<Box<str>>,

}

Geolocation record attached to a Proxy.

All fields are None/Unknown when lookup is disabled or the address is unknown; see GeoData::is_empty.

Examples

use flx::GeoData;

assert!(GeoData::default().is_empty());

Fields
iso_code: Option<Box<str>>

ISO 3166-1 alpha-2 country code (e.g. ID).

name: Option<Box<str>>

Country name.

region_iso_code: Option<Box<str>>

Region ISO code.

region_name: Option<Box<str>>

Region name.

city_name: Option<Box<str>>

City name.

asn: Option<u32>

Autonomous system number.

aso: Option<Box<str>>

Autonomous system organization.

ip_type: IpType

Address classification; see IpType.

continent_code: Option<Box<str>>

Continent code (e.g. AS).

continent_name: Option<Box<str>>

Continent name.

latitude: Option<f64>

Latitude in degrees.

longitude: Option<f64>

Longitude in degrees.

timezone: Option<Box<str>>

IANA timezone (e.g. Asia/Jakarta).

zip: Option<Box<str>>

Postal code.

Implementations

impl GeoData

pub fn is_empty(&self) -> bool

Whether the record carries no lookup data at all.

Used to render geo as {} when geolocation is disabled (or the address is unknown) instead of an object full of nulls.